Move the local cache logic to local_cache_t
